Submission Build
Pushing the build that was submitted for the game jam. There's a glitch where if you hit a fish while the hook is full that fish becomes solid and can never be collected, but the game is still pretty much playable. You added in audio files, but you never got to used. You also hid the volume slider, but the volume title text is still there, which was a mistake.